change from react to static html site
