+++ .eslintrc.js
@@ -7,14 +7,11 @@ module.exports = {
   },
   extends: [
     'airbnb-base',
-    "eslint:recommended",
-    "plugin:react/recommended"
   ],
   globals: {
     Atomics: 'readonly',
     SharedArrayBuffer: 'readonly',
   },
-  "parser": "babel-eslint",
   parserOptions: {
     ecmaVersion: 2018,
     sourceType: 'module',
