--- .github/actions/lhr-out/action.yml
@@ -1,13 +1,0 @@
-name: 'Report Lighthouse'
-description: 'report lighthouse stats for challenge-4'
-inputs:
-  lang:
-    description: 'the programming language in use'
-    required: true
-  server:  
-    description: 'base API endpoint URL'
-    required: true
-    default: 'https://us-central1-buildforsdg.cloudfunctions.net/api'
-runs:
-  using: 'node12'
-  main: 'index.js'
