--- .github/actions/props-out/action.yml
@@ -1,15 +1,0 @@
-name: 'Props Out'
-description: 'expose specific properties in a .properties file'
-inputs:
-  src:
-    description: 'where to look for the .properties file'
-    required: true
-    default: './'
-outputs:
-  frontend: 
-    description: 'URL of the hosted frontend app'
-  backend: 
-    description: 'URL of the hosted backend REST API'
-runs:
-  using: 'node12'
-  main: 'index.js'
