+++ Procfile
@@ -1,1 +1,1 @@
-web: node dist/backend/index.js
+web: node dist/index.js
