+++ app.properties
@@ -1,2 +1,2 @@
-frontend.url = http://example.com
-backend.rest = https://jsonplaceholder.typicode.com/todos
+frontend.url = https://boring-kare-b23aca.netlify.com/
+backend.rest = https://davidshare-covid-19-estimator2.herokuapp.com/api/v1/on-covid-19
