+++ package.json
@@ -7,15 +7,10 @@
   "private": false,
   "scripts": {
     "lint": "eslint ./src/",
-    "start:dev": "nodemon --exec babel-node src/backend/index.js",
-    "start:backend": "yarn build:backend && yarn run serve",
-    "serve": "node dist/backend/index.js",
-    "build:backend": "babel src --out-dir ./dist --source-maps",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject",
-    "predeploy": "yarn run build",
-    "deploy": "gh-pages -d build"
+    "start:dev": "nodemon --exec babel-node src/index.js",
+    "start": "yarn build && yarn run serve",
+    "serve": "node dist/index.js",
+    "build": "babel src --out-dir ./dist --source-maps"
   },
   "dependencies": {
     "@babel/cli": "^7.8.4",
@@ -23,20 +18,16 @@
     "@babel/node": "^7.8.7",
     "@babel/preset-env": "^7.9.0",
     "cors": "^2.8.5",
-    "eslint-plugin-react": "^7.19.0",
     "express": "^4.17.1",
     "gh-pages": "^2.2.0",
-    "node-sass": "^4.13.1",
     "on-covid-19": "^0.1.3",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-scripts": "3.4.0",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.20.1"
+    "eslint-plugin-import": "^2.20.1",
+    "nodemon": "^2.0.3"
   },
   "browserslist": [
     "last 1 Chrome versions"
