+++ public/css/form.css
@@ -1,0 +1,1 @@
+form{padding:40px 20px;display:flex;flex-direction:column}form .input-group{display:flex;flex-direction:column;margin-bottom:20px}form .input-group label{font-size:1em;margin-bottom:5px}form input,form select{background-color:#f9f9f9;font-size:1em;padding:10px;border:0;border:1px solid #f0eeee;border-radius:5px}button{font-size:1em;padding:8px;background-color:#292828;color:#f9f9f9;border:0;border-radius:5px}
