+++ public/css/index.css
@@ -1,0 +1,1 @@
+.header{text-align:center;padding-top:30px}.header img{width:50px;margin-bottom:20px}.header .covid-description-link a{display:block;padding:20px}.main{display:flex;flex-direction:column;width:100%}.main .info-container .info-block{margin-bottom:20px}.main .info-container .info-block h3{margin-bottom:10px}.main .info-container .info-block ul{padding:20px}.main .info-container .info-block ul li{padding:10px 0}.who-info{padding:15px 0}.footer{background-color:#292828;color:#f9f9f9;text-align:center;padding:30px 0}.footer p{margin:10px 0}.footer p a{display:block;padding:10px;color:#f9f9f9}
