+++ public/css/media-min-950.css
@@ -1,0 +1,1 @@
+@media screen and (min-width:950px){.header{padding-bottom:30px}.main{flex-direction:row;justify-content:space-between}.main .info-container{max-width:680px;padding-left:50px;border-left:1px solid #e9e6e6;margin-bottom:30px}.main form{min-width:450px;padding-top:0}}
