+++ public/css/modal.css
@@ -1,0 +1,1 @@
+.modal{display:block;position:fixed;top:0;left:0;width:100%;height:100%;background:rgba(0,0,0,.6)}.modal-content{overflow-y:auto}.modal span.close-icon{border-radius:50px;position:absolute;top:5px;right:5px;font-size:1.5em;height:30px;width:30px;color:#000}.modal .modal-body{position:fixed;background:#fff;width:80%;height:auto;top:50%;left:50%;transform:translate(-50%,-50%)}.show-modal{display:block}.hide-modal{display:none}
