+++ public/css/pseudo-classes.css
@@ -1,0 +1,1 @@
+.modal span.close-icon:hover{cursor:pointer}button:hover{cursor:pointer;color:#292828;border:1px solid #292828;background-color:#f9f9f9}button:focus,input:focus,select:focus{outline:0}
