+++ public/css/reset.css
@@ -1,0 +1,1 @@
+html{box-sizing:border-box;background-color:#fff;font-size:16px;min-width:320px;overflow-x:hidden;overflow-y:auto;text-rendering:optimizeLegibility;text-size-adjust:100%;font-family:Montserrat,sans-serif;color:#292828;transition:.3s}body{-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale}*,::after,::before{box-sizing:inherit;margin:0;padding:0}a{text-decoration:none;color:#333}.container{width:85%;margin:0 auto}
