--- src/frontend/components/App/App.scss
@@ -1,67 +1,0 @@
-.header{
-  text-align: center;
-  padding-top: 30px;
-  
-  img{
-    width: 50px;
-    margin-bottom: 20px;
-  }
-}
-
-.main{
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  .info-container{
-    .info-block{
-      margin-bottom: 20px;
-
-      h3{
-        margin-bottom: 10px;
-      }
-
-      ul{
-        padding: 20px;
-      }
-    }
-  }
-}
-
-.footer{
-  background-color: #626262;
-  color: #f9f9f9;
-  text-align: center;
-  padding: 30px 0;
-
-  p{
-    margin: 5px 0;
-  }
-
-  a{
-    color: #f9f9f9;
-  }
-}
-
-@media screen and (min-width: 950px){
-  .header{
-    padding-bottom: 30px;
-  }
-
-  .main{
-    flex-direction: row;
-    justify-content: space-between;
-
-    .info-container{
-      max-width: 680px;
-      padding-left: 50px;
-      border-left: 1px solid #e9e6e6;
-      margin-bottom: 30px;
-    }
-
-    form{
-      min-width: 450px;
-      padding-top: 0;
-    }
-  }
-}
