--- src/frontend/components/Impact/Impact.scss
@@ -1,27 +1,0 @@
-ul{
-  list-style-type: none;
-
-  .impact-item{
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    width: 100%;
-    justify-content: space-between;
-    margin: 10px 0;
-    background-color: #f9f9f9;
-    padding: 5px 0;
-
-    span{
-      margin: 5px;
-
-      &.description{
-        min-width: 300px;
-      }
-
-      &.value{
-        min-width: 300px;
-        padding-left: 20px;
-      }
-    }
-  }
-}
