--- src/frontend/components/Modal/Modal.scss
@@ -1,42 +1,0 @@
-.modal {
-  display: block;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.6);
-
-  .modal-body {
-    position: fixed;
-    background: white;
-    width: 80%;
-    height: auto;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  span.close-icon{
-    border-radius: 50px;
-    position: absolute;
-    top: 5px;
-    right: 5px;
-    font-size: 1.5em;
-    height: 30px;
-    width: 30px;
-    color: #000;
-
-    &:hover{
-      cursor: pointer;
-    }
-  }
-}
-
-.show-modal {
-  display: block;
-}
-
-.hide-modal {
-  display: none;
-}
