--- src/frontend/reset.scss
@@ -1,47 +1,0 @@
-html {
-  box-sizing: border-box;
-  background-color: #fff;
-  font-size: 16px;
-  min-width: 320px;
-  overflow-x: hidden;
-  overflow-y: auto;
-  text-rendering: optimizeLegibility;
-	text-size-adjust: 100%;
-	font-family: Montserrat, sans-serif;
-	color: #626262;
-	transition: .3s;
-}
-
-body{
-	-webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-*, *::before, *::after {
-	box-sizing: inherit;
-	margin:0;
-	padding:0;
-}
-
-a{
-	text-decoration: none;
-	color: #333;
-}
-
-input:focus,
-button:focus,
-select:focus {
-    outline: none;
-}
-
-.container{
-	width: 85%;
-	margin: 0 auto;
-}
-
-.clear-fix:after {
-  clear: both;
-  content: "";
-  display: table;
-}
-
